requirements.txt
langchain==0.1.0
langchain-ollama==0.1.0
langchain-community==0.0.13
streamlit==1.29.0
python-dotenv==1.0.0
requests==2.31.0